+++ src/felupe/assembly/_axi.py
@@ -114,7 +114,7 @@ class IntegralFormAxisymmetric(IntegralFormCartesian):
     See Also
     --------
     felupe.IntegralForm : Mixed-field integral form container with methods for integration and assembly.
-    felupe.IntegralFormCartesian : Single-field integral form.
+    felupe.assembly.IntegralFormCartesian : Single-field integral form.
 
     """